+++ 24th-App-Team-1-iOS/Core/Extensions/Sources/NotificationName+Extensions.swift
@@ -11,4 +11,5 @@ public extension Notification.Name {
     static let hideTabBar = Notification.Name("hideTabBar")
     static let showTabBar = Notification.Name("showTabBar")
     static let FCMToken = Notification.Name("FCMToken")
+    static let userDidLogin = Notification.Name("userDidLogin")
 }
